+++ src/hooks/useItemList.ts
@@ -1,0 +1,18 @@
+import { useState } from "react";
+import { generateItems } from "../utils";
+
+export function useItemList() {
+  const [items, setItems] = useState(generateItems(1000));
+
+  const addItems = () => {
+    setItems((prevItems) => [
+      ...prevItems,
+      ...generateItems(1000, prevItems.length),
+    ]);
+  };
+
+  return {
+    items,
+    addItems,
+  };
+}
